fix: shared rest functions without server extra
